open: Move MachO.open to macho.rb instead of re-opening the module.
